Final Touches

+++ Assets/Scripts/Scriptable Objects/Player Attacks/RangedPlayerAttack.cs
@@ -9,7 +9,7 @@ public class RangedPlayerAttack : PlayerAttack
 	// cooldown inherited from parent class
 	
 	// for now, range is directly used to set projectile's lifetime
-	public int range; // int because it's used to set PlayerProjectile lifetime
+	public float range; // int because it's used to set PlayerProjectile lifetime
 	public int damage;
 	public float pushStrength;
